Add MIP formulation of CFS
More complicated than MIP formulations of other multivariate filters since
CFS objective involves a fraction, which is not that easy to linearize.

Further, we define k as a variable and pass its value to all FS routines
instead of using a default value of k in all routines (which might go out
of sync once we change it only at some locations).