Add demo of finding alternative subgroup descriptions
Takes a refactored version of the SMT problem for optimal subgroup
discovery and adapts it to formulate the problem of finding alternative
descriptions for a given subgroup. Actually shares a lot of code with the
problem of optimizing WRAcc in the original subgroup. Implements a loop
that finds multiple alternative descriptions on one demo dataset and
conducts a very basic quality/runtime evaluation.